feat: added swift label as a generic component
