+++ Loadify/Model/YouTube/VideoDetails.swift
@@ -53,17 +53,6 @@ struct ErrorModel: Codable {
     var message: String
 }
 
-struct InstagramDetails: Codable, Hashable {
-    
-    let thumbnailURL: String
-    let videoURL: String
-    
-    enum CodingKeys: String, CodingKey {
-        case thumbnailURL = "thumbnail_link"
-        case videoURL = "download_link"
-    }
-}
-
 extension VideoDetails {
     
     static let channelThumbnail = "1https://yt3.ggpht.com/wzh-BL3_M_uugIXZ_ANSSzzBbi_w5XnNSRl4F5DbLAxKdTfXkjgx-kWM1mChdrrMkADRQyB-nQ=s176-c-k-c0x00ffffff-no-rj"
@@ -102,11 +91,3 @@ extension VideoDetails {
     )
 }
 
-extension InstagramDetails {
-    
-    static let thumbnailURL = "https://igcdn.xyz/?token=fa3988c3bdade097ee51a58c8c40437ae447dd17a04819bae63f50623a1b5de2&time=1700717738&file=https%3a%2f%2fscontent.cdninstagram.com%2fv%2ft51.2885-15%2f395265733_364940032536528_5099172351027615466_n.jpg%3fstp%3ddst-jpg_e15%26_nc_ht%3dscontent.cdninstagram.com%26_nc_cat%3d100%26_nc_ohc%3dHC4wuyQisrUAX-2UXkJ%26edm%3dAPs17CUBAAAA%26ccb%3d7-5%26oh%3d00_AfAm5gncIVjqpUJe8F8w_XHRGzEXyQVGtWg3qQk7BpQDeQ%26oe%3d65606889%26_nc_sid%3d10d13b"
-    
-    static let previews: [InstagramDetails] = [
-        InstagramDetails(thumbnailURL: thumbnailURL, videoURL: "")
-    ]
-}
